- Implement root objects' methods with a trait instead of   having them in both Object and Throwable
